fix(cli): speed up shell autocompletions
Speed up shell autocompletions by pulling actual handler code into
a separate entrypoint which is only required when it is going to be
invoked. When testing this speed up shell completions by 6x (from
1.2s to 0.2s [locally on OS X]). The main reason seemed to be e.g.
the React package required by e.g. ink which is deffered into a
separate bundle / entrypoint now. Also `inquirer` seems to have been
quite a burden (used in `init`).
