+++ packages/cdktf-cli/bin/cmds/convert.ts
@@ -1,49 +1,5 @@
 import yargs from "yargs";
-import { convert } from "@cdktf/hcl2cdk";
-import { displayVersionMessage } from "./helper/version-check";
-import { sendTelemetry } from "../../lib/checkpoint";
-import { Errors } from "../../lib/errors";
-import * as fs from "fs-extra";
-import * as path from "path";
-import {
-  readSchema,
-  ConstructsMakerProviderTarget,
-  LANGUAGES,
-  config,
-} from "@cdktf/provider-generator";
-
-function readStreamAsString(stream: typeof process.stdin): Promise<string> {
-  return new Promise((ok, ko) => {
-    if (stream.isTTY) {
-      ko(
-        "No stdin was passed, please use it like this: cat main.tf | cdktf convert > imported.ts"
-      );
-    } else {
-      let string = "";
-      stream.on("data", (data) => (string += data.toString()));
-
-      stream.on("close", () => ok(string));
-      stream.on("error", (err) => ko(err));
-    }
-  });
-}
-
-function findFileAboveCwd(
-  file: string,
-  rootPath = process.cwd()
-): string | null {
-  const fullPath = path.resolve(rootPath, file);
-  if (fs.existsSync(fullPath)) {
-    return fullPath;
-  }
-
-  const parentDir = path.resolve(rootPath, "..");
-  if (fs.existsSync(parentDir) && parentDir !== rootPath) {
-    return findFileAboveCwd(file, parentDir);
-  }
-
-  return null;
-}
+import { requireHandlers } from "./helper/utilities";
 
 class Command implements yargs.CommandModule {
   public readonly command = "convert [OPTIONS]";
@@ -76,39 +32,10 @@ class Command implements yargs.CommandModule {
       })
       .showHelpOnFail(true);
 
-  public async handler({ language }: any) {
-    await displayVersionMessage();
-
-    const providerRequirements: string[] = yargs.argv.provider as string[];
-    const cdktfJsonPath = findFileAboveCwd("cdktf.json");
-    if (cdktfJsonPath) {
-      const cdktfJson = await fs.readJson(cdktfJsonPath);
-      providerRequirements.push(...cdktfJson.terraformProviders);
-    }
-    // Get all the provider schemas
-    const { providerSchema } = await readSchema(
-      providerRequirements.map((spec) =>
-        ConstructsMakerProviderTarget.from(
-          new config.TerraformProviderConstraint(spec),
-          LANGUAGES[0]
-        )
-      )
-    );
-
-    const input = await readStreamAsString(process.stdin);
-    let output;
-    try {
-      const { all, stats } = await convert(input, {
-        language,
-        providerSchema,
-      });
-      output = all;
-      await sendTelemetry("convert", { ...stats, error: false });
-    } catch (err) {
-      throw Errors.Internal("convert", err.message, { language });
-    }
-
-    console.log(output);
+  public async handler(argv: any) {
+    // deferred require to keep cdktf-cli main entrypoint small (e.g. for fast shell completions)
+    const api = requireHandlers();
+    api.convert(argv);
   }
 }
