+++ packages/cdktf-cli/bin/cmds/deploy.ts
@@ -1,11 +1,6 @@
 import * as yargs from "yargs";
-import React from "react";
-import { Deploy } from "./ui/deploy";
 import { config as cfg } from "@cdktf/provider-generator";
-import { renderInk } from "./helper/render-ink";
-import { displayVersionMessage } from "./helper/version-check";
-import { throwIfNotProjectDirectory } from "./helper/check-directory";
-import { checkEnvironment } from "./helper/check-environment";
+import { requireHandlers } from "./helper/utilities";
 
 const config = cfg.readConfigSync();
 
@@ -41,22 +36,9 @@ class Command implements yargs.CommandModule {
       .showHelpOnFail(true);
 
   public async handler(argv: any) {
-    throwIfNotProjectDirectory("deploy");
-    await displayVersionMessage();
-    await checkEnvironment("deploy");
-    const command = argv.app;
-    const outdir = argv.output;
-    const autoApprove = argv.autoApprove;
-    const stack = argv.stack;
-
-    await renderInk(
-      React.createElement(Deploy, {
-        targetDir: outdir,
-        targetStack: stack,
-        synthCommand: command,
-        autoApprove,
-      })
-    );
+    // deferred require to keep cdktf-cli main entrypoint small (e.g. for fast shell completions)
+    const api = requireHandlers();
+    api.deploy(argv);
   }
 }
