+++ packages/cdktf-cli/bin/cmds/destroy.ts
@@ -1,11 +1,6 @@
 import * as yargs from "yargs";
-import React from "react";
-import { Destroy } from "./ui/destroy";
 import { config as cfg } from "@cdktf/provider-generator";
-import { renderInk } from "./helper/render-ink";
-import { displayVersionMessage } from "./helper/version-check";
-import { throwIfNotProjectDirectory } from "./helper/check-directory";
-import { checkEnvironment } from "./helper/check-environment";
+import { requireHandlers } from "./helper/utilities";
 
 const config = cfg.readConfigSync();
 
@@ -40,22 +35,9 @@ class Command implements yargs.CommandModule {
       .showHelpOnFail(true);
 
   public async handler(argv: any) {
-    throwIfNotProjectDirectory("destroy");
-    await displayVersionMessage();
-    await checkEnvironment("destroy");
-    const command = argv.app;
-    const outdir = argv.output;
-    const autoApprove = argv.autoApprove;
-    const stack = argv.stack;
-
-    await renderInk(
-      React.createElement(Destroy, {
-        targetDir: outdir,
-        targetStack: stack,
-        synthCommand: command,
-        autoApprove,
-      })
-    );
+    // deferred require to keep cdktf-cli main entrypoint small (e.g. for fast shell completions)
+    const api = requireHandlers();
+    api.destroy(argv);
   }
 }
