+++ packages/cdktf-cli/bin/cmds/diff.ts
@@ -1,11 +1,6 @@
 import yargs from "yargs";
-import React from "react";
-import { Diff } from "./ui/diff";
 import { config as cfg } from "@cdktf/provider-generator";
-import { renderInk } from "./helper/render-ink";
-import { displayVersionMessage } from "./helper/version-check";
-import { throwIfNotProjectDirectory } from "./helper/check-directory";
-import { checkEnvironment } from "./helper/check-environment";
+import { requireHandlers } from "./helper/utilities";
 
 const config = cfg.readConfigSync();
 
@@ -36,20 +31,9 @@ class Command implements yargs.CommandModule {
       .showHelpOnFail(true);
 
   public async handler(argv: any) {
-    throwIfNotProjectDirectory("diff");
-    await displayVersionMessage();
-    await checkEnvironment("diff");
-    const command = argv.app;
-    const outdir = argv.output;
-    const stack = argv.stack;
-
-    await renderInk(
-      React.createElement(Diff, {
-        targetDir: outdir,
-        targetStack: stack,
-        synthCommand: command,
-      })
-    );
+    // deferred require to keep cdktf-cli main entrypoint small (e.g. for fast shell completions)
+    const api = requireHandlers();
+    api.diff(argv);
   }
 }
