+++ packages/cdktf-cli/bin/cmds/get.ts
@@ -1,18 +1,7 @@
 import yargs from "yargs";
-import React from "react";
-import { Language, LANGUAGES, config as cfg } from "@cdktf/provider-generator";
-import { Get } from "./ui/get";
-import { renderInk } from "./helper/render-ink";
-import { displayVersionMessage } from "./helper/version-check";
-import { throwIfNotProjectDirectory } from "./helper/check-directory";
-import { checkEnvironment } from "./helper/check-environment";
-
+import { LANGUAGES, config as cfg } from "@cdktf/provider-generator";
+import { requireHandlers } from "./helper/utilities";
 const config = cfg.readConfigSync();
-
-interface Arguments {
-  output: string;
-  language: Language;
-}
 
 class Command implements yargs.CommandModule {
   public readonly command = "get [OPTIONS]";
@@ -38,33 +27,9 @@ class Command implements yargs.CommandModule {
       });
 
   public async handler(argv: any) {
-    throwIfNotProjectDirectory("get");
-    await displayVersionMessage();
-    await checkEnvironment("get");
-    const args = argv as Arguments;
-    const providers = config.terraformProviders ?? [];
-    const modules = config.terraformModules ?? [];
-    const { output, language } = args;
-
-    const constraints: cfg.TerraformDependencyConstraint[] = [
-      ...providers,
-      ...modules,
-    ];
-
-    if (constraints.length === 0) {
-      console.error(
-        `ERROR: Please specify providers or modules in "cdktf.json" config file`
-      );
-      process.exit(1);
-    }
-
-    await renderInk(
-      React.createElement(Get, {
-        codeMakerOutput: output,
-        language: language,
-        constraints,
-      })
-    );
+    // deferred require to keep cdktf-cli main entrypoint small (e.g. for fast shell completions)
+    const api = requireHandlers();
+    api.get(argv);
   }
 }
