+++ packages/cdktf-cli/bin/cmds/helper/init-templates.ts
@@ -1,0 +1,12 @@
+import * as fs from "fs-extra";
+import * as path from "path";
+import { projectRootPath } from "./utilities";
+
+export const templatesDir = path.join(projectRootPath(), "templates");
+const availableTemplates = fs
+  .readdirSync(templatesDir)
+  .filter((x) => !x.startsWith("."));
+export const templates: string[] = [];
+for (const template of availableTemplates) {
+  templates.push(template);
+}
