+++ packages/cdktf-cli/bin/cmds/login.ts
@@ -1,11 +1,5 @@
 import yargs from "yargs";
-import { TerraformLogin } from "./helper/terraform-login";
-import * as terraformCloudClient from "./helper/terraform-cloud-client";
-import * as chalk from "chalk";
-import { terraformCheck } from "./helper/terraform-check";
-import { displayVersionMessage } from "./helper/version-check";
-
-const chalkColour = new chalk.Instance();
+import { requireHandlers } from "./helper/utilities";
 
 class Command implements yargs.CommandModule {
   public readonly command = "login";
@@ -14,41 +8,9 @@ class Command implements yargs.CommandModule {
   public readonly builder = (args: yargs.Argv) => args.showHelpOnFail(true);
 
   public async handler(argv: any) {
-    await terraformCheck();
-    await displayVersionMessage();
-
-    const args = argv as yargs.Arguments;
-    if (args["_"].length > 1) {
-      console.error(
-        chalkColour`{redBright ERROR: 'cdktf login' command cannot have more than one argument.}\n`
-      );
-      yargs.showHelp();
-      process.exit(1);
-    }
-
-    const terraformLogin = new TerraformLogin();
-    const token = await terraformLogin.askToLogin();
-    if (token == "") {
-      console.error(
-        chalkColour`{redBright ERROR: couldn't configure Terraform Cloud credentials.}\n`
-      );
-      process.exit(1);
-    }
-
-    // Get user details if token is set
-    const userAccount = await terraformCloudClient.getAccountDetails(token);
-    if (userAccount) {
-      const username = userAccount.data.attributes.username;
-      console.log(
-        chalkColour`\n{greenBright cdktf has successfully configured Terraform Cloud credentials!}`
-      );
-      console.log(chalkColour`\nWelcome {bold ${username}}!`);
-    } else {
-      console.error(
-        chalkColour`{redBright ERROR: couldn't configure Terraform Cloud credentials.}\n`
-      );
-      process.exit(1);
-    }
+    // deferred require to keep cdktf-cli main entrypoint small (e.g. for fast shell completions)
+    const api = requireHandlers();
+    api.login(argv);
   }
 }
