+++ packages/cdktf-cli/package.json
@@ -77,7 +77,6 @@
     ]
   },
   "devDependencies": {
-    "jsii-srcmak": "^0.1.448",
     "@apollo/client": "^3.5.7",
     "@cdktf/provider-generator": "0.0.0",
     "@graphql-tools/graphql-file-loader": "^6.2.7",
@@ -134,6 +133,7 @@
     "ink-use-stdout-dimensions": "^1.0.5",
     "inquirer": "^8.2.0",
     "jest": "^26.6.3",
+    "jsii-srcmak": "^0.1.448",
     "lodash.isequal": "^4.5.0",
     "log4js": "^6.3.0",
     "nock": "^13.2.2",
